added basic editorState logic to hooks
